Move the `javascript` package inside `linker.backend`.
Only the backends should be concerned with manipulating JavaScript
trees and printing them.
