+++ tools/shared/src/main/scala/org/scalajs/core/tools/linker/backend/BasicLinkerBackend.scala
@@ -15,7 +15,10 @@ import org.scalajs.core.tools.io.WritableVirtualJSFile
 import org.scalajs.core.tools.linker.LinkingUnit
 import org.scalajs.core.tools.linker.analyzer.SymbolRequirement
 import org.scalajs.core.tools.linker.backend.emitter.Emitter
-import org.scalajs.core.tools.javascript.{JSFileBuilder, JSFileBuilderWithSourceMap}
+
+import org.scalajs.core.tools.linker.backend.javascript.{
+  JSFileBuilder, JSFileBuilderWithSourceMap
+}
 
 /** The basic backend for the Scala.js linker.
  *
