+++ tools/shared/src/main/scala/org/scalajs/core/tools/linker/backend/emitter/FunctionEmitter.scala
@@ -24,8 +24,8 @@ import org.scalajs.core.tools.sem._
 import CheckedBehavior._
 
 import org.scalajs.core.tools.linker.LinkedClass
-import org.scalajs.core.tools.linker.backend.OutputMode
-import org.scalajs.core.tools.javascript.{Trees => js}
+import org.scalajs.core.tools.linker.standard.OutputMode
+import org.scalajs.core.tools.linker.backend.javascript.{Trees => js}
 
 import java.io.StringWriter
