+++ tools/shared/src/main/scala/org/scalajs/core/tools/linker/backend/javascript/JSBuilders.scala
@@ -7,7 +7,7 @@
 \*                                                                      */
 
 
-package org.scalajs.core.tools.javascript
+package org.scalajs.core.tools.linker.backend.javascript
 
 import scala.annotation.tailrec
